refactor: cc.crss.alcoholonly -> chaos.alcohol
