--- src/main/java/cc/crss/alcoholonly/Item/custom/MolotovItem.java
@@ -1,36 +1,0 @@
-package cc.crss.alcoholonly.Item.custom;
-
-import cc.crss.alcoholonly.entity.custom.Molotov_BottleEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.sound.SoundEvents;
-import net.minecraft.stat.Stats;
-import net.minecraft.util.Hand;
-import net.minecraft.util.TypedActionResult;
-import net.minecraft.world.World;
-
-public class MolotovItem extends Item {
-    public MolotovItem(Settings settings) {
-        super(settings);
-    }
-
-    public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
-        ItemStack itemStack = user.getStackInHand(hand); // creates a new ItemStack instance of the user's itemStack in-hand
-        world.playSound(null, user.getX(), user.getY(), user.getZ(), SoundEvents.ENTITY_SNOWBALL_THROW, SoundCategory.NEUTRAL, 0.5F, 1F); // plays a globalSoundEvent
-        if (!world.isClient) {
-            Molotov_BottleEntity entity = new Molotov_BottleEntity(user, world);
-            entity.setItem(itemStack);
-            entity.setVelocity(user, user.getPitch(), user.getYaw(), 0.0F, 1.5F, 1.0F);
-            world.spawnEntity(entity); // spawns entity
-        }
-
-        user.incrementStat(Stats.USED.getOrCreateStat(this));
-        if (!user.getAbilities().creativeMode) {
-            itemStack.decrement(1); // decrements itemStack if user is not in creative mode
-        }
-
-        return TypedActionResult.success(itemStack, world.isClient());
-    }
-}
